Remove fireant dataframe-level pagination but keep sort functionality
