+++ fireant/queries/sorting.py
@@ -1,18 +1,13 @@
+from typing import Tuple
+
 import pandas as pd
 from pandas.core.dtypes.common import is_datetime64_ns_dtype
+from pypika import Order
 
 from fireant.utils import alias_selector
 
-from pypika import Order
 
-
-def _get_window(limit, offset):
-    start = offset
-    end = offset + limit if None not in (offset, limit) else limit
-    return start, end
-
-
-def _apply_sorting(orders):
+def _get_sorting_schema(orders) -> Tuple[list, bool]:
     sort_values, ascending = zip(
         *[
             (alias_selector(field.alias), orientation is Order.asc)
@@ -22,54 +17,44 @@ def _apply_sorting(orders):
     return list(sort_values), ascending
 
 
-def paginate(data_frame, widgets, orders=(), limit=None, offset=None):
+def sort(data_frame, widgets, orders=()):
     """
     :param data_frame:
-        The result set to paginate.
+        The result set to sort.
     :param widgets:
-        An iterable of widgets that the pagination is being applied for.
+        An iterable of widgets that the sort is being applied for.
     :param orders:
         An iterable of (<Dimension/Metric>, pypika.Order)
-    :param limit:
-        A limit of the number of data points/series
-    :param offset:
-        A offset of the number of data points/series
     :return:
-        A paginated data frame. If the widget required grouped pagination, then there should be an upperbound
-        `limit*(n_index_level_0)`. Otherwise the data frame should have the same length as the limit.
+        A sorted data frame.
     """
     if len(data_frame) == 0:
         return data_frame
 
-    start, end = _get_window(limit, offset)
-    group_pagination = isinstance(data_frame.index, pd.MultiIndex) and any(
-        [getattr(widget, "group_pagination", False) for widget in widgets]
+    needs_group_sort = isinstance(data_frame.index, pd.MultiIndex) and any(
+        [getattr(widget, 'group_sort', False) for widget in widgets]
     )
 
-    if group_pagination:
-        return _group_paginate(data_frame, start, end, orders)
-    return _simple_paginate(data_frame, start, end, orders)
+    if needs_group_sort:
+        return _group_sort(data_frame, orders)
+    return _simple_sort(data_frame, orders)
 
 
-def _simple_paginate(data_frame, start=None, end=None, orders=()):
+def _simple_sort(data_frame, orders=()):
     """
-    Applies pagination which limits the number of rows in the data frame.
+    Sorts the data frame.
 
     :param data_frame:
-        A data frame to paginate
-    :param start:
-        The index starting point to slice the data frame at
-    :param end:
-        The index ending point to slice the data frame at
+        A data frame to sort
     :param orders:
         A list of tuples that contain a slicer field definition (with an alias matching the columns of the data frame)
         and a pypika.Order.
     """
     if orders:
-        sort, ascending = _apply_sorting(orders)
+        sort, ascending = _get_sorting_schema(orders)
         data_frame = data_frame.sort_values(by=sort, ascending=ascending)
 
-    return data_frame[start:end]
+    return data_frame
 
 
 def _index_isnull(data_frame):
@@ -90,17 +75,12 @@ def _aggregate_dimension_groups(group):
     return group.sum()
 
 
-def _group_paginate(data_frame, start=None, end=None, orders=()):
+def _group_sort(data_frame, orders=()):
     """
-    Applies pagination which limits the number of rows in the data frame grouped by the zeroth index level. This will
-    in turn paginate the number of series in the data frame.
+    Applies sorting for a number of series in a dataframe
 
     :param data_frame:
-        A data frame to paginate
-    :param start:
-        The index starting point to slice the data frame at
-    :param end:
-        The index ending point to slice the data frame at
+        A data frame to order
     :param orders:
         A list of tuples that contain a slicer field definition (with an alias matching the columns of the data frame)
         and a pypika.Order.
@@ -118,14 +98,12 @@ def _group_paginate(data_frame, start=None, end=None, orders=()):
 
     if orders:
         aggregated_df = dimension_groups.aggregate(_aggregate_dimension_groups)
-        sort, ascending = _apply_sorting(orders)
+        sort, ascending = _get_sorting_schema(orders)
         sorted_df = aggregated_df.sort_values(by=sort, ascending=ascending)
-        sorted_dimension_values = tuple(sorted_df.index)[start:end]
+        sorted_dimension_values = tuple(sorted_df.index)
 
     else:
-        sorted_dimension_values = tuple(dimension_groups.apply(lambda g: g.name))[
-            start:end
-        ]
+        sorted_dimension_values = tuple(dimension_groups.apply(lambda g: g.name))
 
     sorted_dimension_values = (
         pd.Index(sorted_dimension_values, name=dimension_levels[0])
@@ -133,7 +111,7 @@ def _group_paginate(data_frame, start=None, end=None, orders=()):
         else pd.MultiIndex.from_tuples(sorted_dimension_values, names=dimension_levels)
     )
 
-    def _apply_pagination(df):
+    def _apply_sort(df):
         # This function applies sorting by using the sorted dimension values as an index to select values in the right
         # order out of the data frame. The index must be filtered to only values that are in this data frame, since
         # there might be missing combinations of index values.
@@ -156,5 +134,5 @@ def _group_paginate(data_frame, start=None, end=None, orders=()):
     return (
         data_frame.sort_values(data_frame.index.names[0], ascending=True)
         .groupby(level=0)
-        .apply(_apply_pagination)
+        .apply(_apply_sort)
     )
