+++ fireant/tests/queries/test_fetch_operations.py
@@ -14,7 +14,7 @@ from fireant.utils import alias_selector
 
 
 @patch('fireant.queries.builder.dataset_query_builder.scrub_totals_from_share_results', side_effect=lambda *args, **kwargs: args[0])
-@patch('fireant.queries.builder.dataset_query_builder.paginate', side_effect=lambda *args, **kwargs: args[0])
+@patch('fireant.queries.builder.dataset_query_builder.sort', side_effect=lambda *args, **kwargs: args[0])
 @patch('fireant.queries.builder.dataset_query_builder.fetch_data')
 class QueryBuilderOperationsTests(TestCase):
     def test_operations_evaluated(self, mock_fetch_data: Mock, *mocks):
