+++ fireant/widgets/base.py
@@ -61,9 +61,9 @@ class Widget:
 
 
 class TransformableWidget(Widget):
-    # This attribute can be overridden in order to paginate in groups. Useful in cases like for charts where pagination
+    # This attribute can be overridden in order to sort groups. Useful in cases like for charts where sorting
     # should be applied to the number of series rather than the number of data points.
-    group_pagination = False
+    group_sort = False
 
     @staticmethod
     def hide_data_frame_indexes(data_frame, dimensions_to_hide):
